Provide way to simulate network issues for tests
This introduces a proxy that simulates different network issues and a
client to manage it. That is handy to be able to test for issues that
can only be reproduced in specific scenarios/environments.

The management API provided here is slimmed down to only reproduce peer
resets but can be further extended if/when necessary.

More information: https://github.com/shopify/toxiproxy